requirements.txt
microphone==0.0.1
pluginmanager==0.3.2
pocketsphinx==0.0.9
PyAudio==0.2.9
pyzmq==15.2.0
SpeechRecognition==3.3.1
wheel==0.26.0
